adds nohint option to rules
The window manager adheres to application desktop hints. An application
may set this hint to have the window manager spawn its window on a
particular workspace (by index). Some programs use this to spawn their
window on the workspace it was last on before being minimized to the
system tray. Sometimes, this is unwanted behavior.  To counter this, the
user can set a rule for a particular application to ignore desktop hints
altogether.