circumvents X server race condition
When firing off a bunch of UnmapWindow requests in rapid succession (by,
for example, continually calling toggle_workspace), a race condition in
the X server would occur. This race condition disrupted proper client
handling and even caused client windows to be destroyed by the X server.
This commit implements a circumvention to the problem.