fixes non-wm client move offsetting
Initially, when moving a window by clicking and dragging a
move-hooked area within the client application, its position
would not be remembered when subsequently resizing that
client by clicking and dragging a resize-hooked area
of the application. This would result in jumping of the
client application's position. This commit fixes this
issue.