has clients wedge to screen on root size change
Previously, whenever the root window was resized (using `xrandr`, for
instance), clients would retain their position relative to the origin of
the root window, even if the screen size became so small that those
windows would no longer be visible on screen. This commit has clients
wedged within the boundaries of the screen whenever the root window's
size changes.