adds focus conditional to mousebinds
This commit implements a conditional focus specifier to mousebinds.
When executing mousebinds, kranewm config branches on whether the
mousebind was performed on a client or on the root window. When
performed on a client, kranewm would previously default to focusing
that client. Some mousebinds should not behave this way. For
instance, a mousebind that activates the next workspace, performed
when hovering the cursor over a client, should not focus that client.
Before this commit, kranewm did move focus to the client in question.

This commit also adds mousebinds (both client-bound and root-bound)
for moving focus forward and backward through the focus cycle.