changes mousebinds to operate on a target window
Previously, a mousebind would either operate on a client window, or
not (in which case it would operate on the root window). To make
mousebinds more flexible, such as to allow them to, for instance,
operate on the sidebar, a value was added to them to specify on
which target they operate.